% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{change2date}
\alias{change2date}
\title{Detection and casting for date columns}
\usage{
change2date(data)
}
\arguments{
\item{data}{The tibble to convert}
}
\value{
A mutated tibble
}
\description{
\code{change2date} internal function
}
