% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{change2yn}
\alias{change2yn}
\title{Detection and casting for yes/no columns}
\usage{
change2yn(data)
}
\arguments{
\item{data}{The tibble to convert}
}
\value{
A mutated tibble
}
\description{
\code{change2yn} internal function
}
