% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{detect_date}
\alias{detect_date}
\title{Detection and casting for columns containing DD.MM.YYYY}
\usage{
detect_date(column)
}
\arguments{
\item{column}{A vector (typically a column from an imported sheet)}
}
\value{
A vector of type date if suitable, original structure otherwise
}
\description{
\code{detect_date} is an internal function, taking a vector as input
and returning either original or casted vector
}
