% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{detect_numeric}
\alias{detect_numeric}
\title{Detection of numeric variables
Tests if all entries can be casted to numeric}
\usage{
detect_numeric(column)
}
\arguments{
\item{column}{A vector (typically a column from an imported sheet)}
}
\value{
A vector of type numeric if suitable, original structure otherwise
}
\description{
\code{detect_numeric} internal function
}
