% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{import_excelsheets}
\alias{import_excelsheets}
\title{Import (selected?) sheets from Excelfile}
\usage{
import_excelsheets(
  file,
  sheets = NULL,
  skip_empty_cols = FALSE,
  skip_codecols = TRUE
)
}
\arguments{
\item{file}{The path to the Excelfile}

\item{sheets}{A vector with sheet numbers to include,
ignored if NULL (the default)}

\item{skip_empty_cols}{Logical, if FALSE (default), empty columns are retained}

\item{skip_codecols}{Logical, if TRUE (default), columns with numeric level for categories are removed}
}
\value{
A named list with tibbles for imported sheets
}
\description{
\code{import_exelsheets} imports all or selected sheets
}
