% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{sheet_summary}
\alias{sheet_summary}
\title{Create summary for all imported sheets}
\usage{
sheet_summary(data, exclude = c("Nr", "PatientStudyID", "MonthAndYearOfBirth"))
}
\arguments{
\item{data}{List of tibbles}

\item{exclude}{Vector of colum names not to be summarized,
by default Nr, PatientStudyID, MonthAndYearOfBirth}
}
\value{
List with entries per sheet
}
\description{
\code{sheet_summary} creates suitable summaries for all columns
}
