% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{show_excelfiles}
\alias{show_excelfiles}
\title{Show Excelfiles in selected folder}
\usage{
show_excelfiles(folder = "Data", pattern = "")
}
\arguments{
\item{folder}{The folder to check}

\item{pattern}{optional search pattern to narrow down search}
}
\value{
A vector with full paths
}
\description{
\code{show_excelfiles} internal function
}
