% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{show_excelsheets}
\alias{show_excelsheets}
\title{\code{show_excelsheets} lists all sheets found in the selected Excelfile}
\usage{
show_excelsheets(file)
}
\arguments{
\item{file}{Path to Excelfile}
}
\value{
A vector with sheetnames
}
\description{
\code{show_excelsheets} lists all sheets found in the selected Excelfile
}
