% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{change2numeric}
\alias{change2numeric}
\title{Detection and casting for numeric columns}
\usage{
change2numeric(data)
}
\arguments{
\item{data}{The tibble to convert}
}
\value{
A mutated tibble
}
\description{
\code{change2numeric} internal function
}
